Bump akka version
