+++ build.sbt
@@ -1,6 +1,6 @@
 name := "brando"
 organization := "com.digital-achiever"
-version := "3.1.13"
+version := "3.1.14"
 scalaVersion := "2.12.6"
 scalacOptions ++= Seq("-unchecked", "-deprecation", "-feature")
 
@@ -17,9 +17,9 @@ credentials += Credentials(
 )
 
 libraryDependencies ++= Seq(
-  "com.typesafe.akka" %% "akka-actor" % "2.5.13",
+  "com.typesafe.akka" %% "akka-actor" % "2.6.13",
   "org.scalatest" %% "scalatest" % "3.0.5" % "test",
-  "com.typesafe.akka" %% "akka-testkit" % "2.5.13" % "test"
+  "com.typesafe.akka" %% "akka-testkit" % "2.6.13" % "test"
 )
 
 parallelExecution in Test := false
